test(engine): add enrichment tests for kernelized shift guard and semantic dark matter engine
Add ~190 lines of unit tests to each module covering custom config
construction, shift/promotion detection, expansion triggering, board
freshness/state, content hash, epoch propagation, summary counters,
reset, history, error Display impls, specimen family coverage, and
corpus evidence hash uniqueness. Fix minor style issues in neighboring
enrichment integration tests.

+++ crates/franken-engine/tests/catastrophe_witness_generator_enrichment_integration.rs
@@ -470,7 +470,7 @@ fn minimize_witness_already_minimal() {
 fn minimize_witness_reduction_ratio() {
     let w = make_witness("min-3", 800_000, 200_000);
     let result = minimize_witness(&w).unwrap();
-    let ratio = result.reduction_ratio_millionths(w.replay_steps as u64);
+    let ratio = result.reduction_ratio_millionths(w.replay_steps);
     assert!(ratio > 0, "should have positive reduction ratio");
     assert!(ratio <= MILLIONTHS, "ratio should not exceed 1.0");
 }
